==> processing_scripts/3p/01-ambient_RNA_check.runfile.Rmd <==
---
title: 'scRNA QC'
author: "FHIL\nDerrikGratz"
date: '`r Sys.Date()`'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 

library(tidyverse)  ## General R logic
library(here)       ## Easier specification of file locations
library(yaml)       ## parses config yaml
library(Seurat)
library(SoupX)
library(reshape2)   ## DF manipulation
library(data.table) ## rbindlist
# library(FHIL.SC.utils) ## Custom functions
set.seed(33)
```

```{r}
figures <- list()
```

# Functions

```{r}
# Scale somehow outputs clustering information for cells not present in the input filtered counts matrix. So technically the contamination estimate is on a slight subset of the full input data. Since we're not planning to correct the counts, but only
# report the contamination estimate, I'm not too worried about this. It does seem problematic in general, but I plan to read in
# the filtered matrix and work from there, not rely on their processing. 
runSoupX <- function(tod,
                     toc,
                     clusters,
                     doPlot = TRUE,
                     roundToInt = TRUE){
  ## this is intended for scale data. They have outputs with additional barcodes somehow
  if (!all(colnames(toc) %in% names(clusters)) | !(all(names(clusters) %in% colnames(toc)))){
    warning('Names of clusters and filtered data to not match. Subsetting to values present in both')
    toc_full <- toc
    toc <- toc[,colnames(toc)[colnames(toc) %in% names(clusters)]]
    clusters <- clusters[names(clusters)[names(clusters) %in% colnames(toc)]]
    store_full = TRUE
  } else {
    store_full = FALSE
  }
  sc <- SoupX::SoupChannel(tod = tod,
                           toc = toc[,names(clusters)])
  sc <- SoupX::setClusters(sc, clusters)
  sc <- SoupX::autoEstCont(sc, doPlot = doPlot, contaminationRange = c(0.001,0.8))
  sc$adjusted_counts <- SoupX::adjustCounts(sc, roundToInt = roundToInt)
  sc$plot <- grDevices::recordPlot()
  
  if (store_full) {
    sc$toc_full <- toc_full
  }
  sc
}
```

# Load config

```{r}
source(here('config/3p/generate_metadata.R'))
metadata <- read.csv(here('config/3p/metadata.csv'))
metadata
```

# Load data

```{r}
objs <- list()
```

## flex

```{r}
kit <- 'flex'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], 'per_sample_outs', file, 'count/sample_filtered_feature_bc_matrix.h5')
  filtered_counts <- Read10X_h5(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], 'per_sample_outs', file, 'count/sample_raw_feature_bc_matrix.h5')
  unfiltered_counts <- Read10X_h5(unfiltered_counts)
  unfiltered_counts <- unfiltered_counts[rownames(filtered_counts),]
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], 'count/analysis_csv/clustering/gene_expression_kmeans_10_clusters/clusters.csv')
  clusters <- read.csv(clusters, header=TRUE)
  clusters <- setNames(as.list(clusters$Cluster), clusters$Barcode)
  
  clusters <- clusters[colnames(filtered_counts)]
  
  
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## fluent_v4

```{r}
kit <- 'fluent_v4'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  sensitivity <- samplesheets[[kit]][i,'Sensitivity']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'filtered_matrix', paste0('sensitivity_', sensitivity))
  filtered_counts <- Read10X(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'raw_matrix')
  unfiltered_counts <- Read10X(unfiltered_counts)
  
  barcodes <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'raw_matrix', 'barcodes.tsv.gz')
  barcodes <- read.table(gzfile(barcodes))
  barcodes$cell_id <- 1:nrow(barcodes)
  
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'clustering', paste0('sensitivity_', sensitivity), 'clusters.csv')
  clusters <- read.csv(clusters)
  clusters <- right_join(barcodes, clusters, by ='cell_id')
  clusters <- setNames(as.list(clusters$graph), clusters$V1)
  
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## fluent_V

```{r}
kit <- 'fluent_V'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  sensitivity <- samplesheets[[kit]][i,'Sensitivity']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'filtered_matrix', paste0('sensitivity_', sensitivity))
  filtered_counts <- Read10X(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'raw_matrix')
  unfiltered_counts <- Read10X(unfiltered_counts)
  
  barcodes <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'raw_matrix', 'barcodes.tsv.gz')
  barcodes <- read.table(gzfile(barcodes))
  barcodes$cell_id <- 1:nrow(barcodes)
  
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'clustering', paste0('sensitivity_', sensitivity), 'clusters.csv')
  clusters <- read.csv(clusters)
  clusters <- right_join(barcodes, clusters, by ='cell_id')
  clusters <- setNames(as.list(clusters$graph), clusters$V1)
  
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## gemx_3p

```{r}
kit <- 'gemx_3p'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                               'outs', 'per_sample_outs', file, 'count/sample_filtered_feature_bc_matrix.h5')
  filtered_counts <- Read10X_h5(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                                 'outs', 'multi/count/raw_feature_bc_matrix.h5')
  unfiltered_counts <- Read10X_h5(unfiltered_counts)
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'outs', 'per_sample_outs',
                        file, 'count/analysis/clustering/gene_expression_graphclust/clusters.csv')
  clusters <- read.csv(clusters)
  clusters <- setNames(as.list(clusters$Cluster), clusters$Barcode)
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## nextgem_3p

```{r}
kit <- 'nextgem_3p'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                               'outs', 'per_sample_outs', file, 'count/sample_filtered_feature_bc_matrix.h5')
  filtered_counts <- Read10X_h5(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                                 'outs', 'multi/count/raw_feature_bc_matrix.h5')
  unfiltered_counts <- Read10X_h5(unfiltered_counts)
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'outs', 'per_sample_outs',
                        file, 'count/analysis/clustering/gene_expression_graphclust/clusters.csv')
  clusters <- read.csv(clusters)
  clusters <- setNames(as.list(clusters$Cluster), clusters$Barcode)
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## parse_v3

```{r}
kit <- 'parse_v3'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'DGE_filtered')
  filtered_counts <- ReadParseBio(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'DGE_unfiltered')
  unfiltered_counts <- ReadParseBio(unfiltered_counts)
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'report/cluster_assignment.csv')
  clusters <- read.csv(clusters)
  clusters <- setNames(as.list(clusters$cluster), clusters$bc_wells)
  
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## scale

```{r}
kit <- 'scale'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], #'scale/downsampled_runs/ScaleRNA_no_multimapping',
                               'samples',
                               paste0(file, '.downsampled.filtered.matrix'))
  filtered_counts <- Read10X(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]],
                                 'samples',
                                 paste0(file, '.downsampled.unfiltered.matrix'))
  unfiltered_counts <- Read10X(unfiltered_counts)
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]],
                        'cellTyping',
                        paste0(file, '.downsampled'),
                        paste0(file, '.downsampled_seurat_clustering_results.csv'))
  
  clusters <- read.csv(clusters)
  clusters <- setNames(as.list(clusters$seurat_clusters), gsub('(.+)_F.+', '\\1', clusters$cellBarcode))
  
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## bd

```{r, eval=FALSE}
kit <- 'bd'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  individual <- samplesheets[[kit]][i,'Individual']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]],
                               individual, paste0('BD2-WTA-downsampled_', individual, '_RSEC_MolsPerCell_MEX'))
  filtered_counts <- Read10X(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]],
                                 individual, paste0('BD2-WTA-downsampled_', individual, '_RSEC_MolsPerCell_Unfiltered_MEX'))
  unfiltered_counts <- Read10X(unfiltered_counts)
  unfiltered_counts <- unfiltered_counts[rownames(filtered_counts), ]
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], individual,
                        paste0('BD2-WTA-downsampled_', individual, '_Seurat.rds'))
  
  clusters <- readRDS(clusters)
  clusters <- clusters$Cell_Type_Experimental
  # clusters <- setNames(as.list(clusters$seurat_clusters), gsub('(.+)_F.+', '\\1', clusters$cellBarcode))
  
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

# SoupX

```{r}
objs <- lapply(objs, function(x) {
  runSoupX(x$tod, x$toc, x$clusters, doPlot = TRUE, roundToInt = TRUE)
})
```

# Save RDS

```{r, eval=FALSE}
saveRDS(objs, here('rds/01-soup_channels.rds'), compress=FALSE)
```


```{r, eval=TRUE}
objs <- readRDS(here('rds/01-soup_channels.rds'))
```

## Append individual reruns

```{r, eval=FALSE}
objs_rerun <- objs
objs <- readRDS(here('rds/01-soup_channels.rds'))
for (x in names(objs_rerun)) {
  objs[[x]] <- objs_rerun[[x]]
}
saveRDS(objs, here('rds/01-soup_channels.rds'), compress=FALSE)
```

# Summarize contamination

```{r}
extractSoupXContamEst <- function(sc){
  rho <- sc$fit$rhoEst
  rho_low <- sc$fit$rhoFWHM[1]
  rho_high <- sc$fit$rhoFWHM[2]
  return(list(rho_low = rho_low, rho = rho, rho_high = rho_high))
}
soupx_results <- data.table::rbindlist(lapply(objs, extractSoupXContamEst), idcol = 'Sample')
soupx_results
```

```{r}
figures[['rho_barchart']] <-
soupx_results %>%
  # reshape2::melt() %>%
  merge(metadata, by = 'Sample') %>%
  ggplot(aes(x=paste0(Individual, Replicate), y=rho)) +
  geom_col() +
  geom_errorbar(aes(ymin=rho_low, ymax=rho_high), width=0.4) +
  facet_wrap(~ Kit, scales='free_x', ncol=4) +
  labs(x='Sample', y='Contamination fraction')
figures[['rho_barchart']] 
```

# Correct counts (if desired)

```{r, eval = FALSE}
objs <- lapply(objs, function(x) x$adjusted_counts)
```

```{r, eval = TRUE}
objs <- lapply(objs, function(x) {
  if (!is.null(x$toc_full)) {
    x$toc_full
  } else {
    x$toc
}})
```

# Save objs

```{r}
objs <- lapply(objs, CreateSeuratObject)
for (obj in names(objs)){
  Project(objs[[obj]]) <- obj
}
for (obj in names(objs)){
  objs[[obj]]$orig.ident <- obj
}
```

```{r}
saveRDS(objs, here('rds/3p/01_raw_objs.rds'), compress=TRUE)
```

# Render report

```{r}
## This only works if all the variables called in the format file 
## are in the current environment. Check the format file to tweak
## aesthetics 
rmarkdown::render(here('processing_scripts/3p/01-ambient_RNA_check.format.Rmd'),
                  output_file = '01_soupx.html',
                  output_dir = here('reports/3p'))
```
