==> processing_scripts/3p/05.1-sc3s.ipynb <==
import scanpy as sc
import sc3s
from pyprojroot import here

here()

adata = sc.read_h5ad(here('rds/3p/sc3_anndata/Flex.h5ad'))

adata

# remove lowly expressed genes
# sc.pp.filter_genes(adata, min_cells=10)

# # log transform the data
# sc.pp.log1p(adata)

# # dimensionality reduction with PCA
# sc.tl.pca(adata, svd_solver='arpack')
adata.obsm["X_pca"] = adata.obsm["X_PCA_allgenes"]


sc3s.tl.consensus(adata, n_clusters=list(range(5,30,1)))

adata.obs['sc3s_28']   # returns the cluster labels
