==> processing_scripts/3p/06-composition.runfile.Rmd <==
---
title: 'SC processing'
author: "FHIL\nDerrikGratz"
date: '`r Sys.Date()`'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 

library(tidyverse)  ## General R logic
library(here)       ## Easier specification of file locations
library(yaml)       ## parses config yaml
library(Seurat)
library(reshape2)   ## DF manipulation
# library(clustree)   ## Resolution selection for clustering
# library(ggraph)     ## To render clustree
# library(ggforestplot) ## Striped lines for dotplots
# library(SingleCellExperiment)
# library(gridExtra)  ## arrange plots
library(janitor) ## Tabyl, adorn_rounding
# library(ggrastr) ## Rasterize images for smaller size
library(DT)     ## Datatables
rbindlist <- data.table::rbindlist
set.seed(33)
```

# Functions

```{r}
coarse_mapping <- c(
   'CD8+ T' = 'T',
   'CD4+ T' = 'T',
   'T' = 'T',
   'NK' = 'NK',
   'B naive' = 'B',
   'B memory' = 'B',
   'B' = 'B',
   'Classical monocyte' = 'Monocyte',
   'Non-classical monocyte' = 'Monocyte',
   'Monocyte' = 'Monocyte',
   'Megakaryocyte' = 'Megakaryocyte',
   'Dendritic' = 'Dendritic',
   'pDC' = 'pDC',
   'Granulocyte' = 'Granulocyte',
   'Unknown' = 'Unknown'
)
```


# Load data

```{r}
objs <- readRDS(here('rds/3p/05_merged_objs_post_clustering.rds'))
kit_order <- read.table(here('config/3p/kit_order.txt'))$V1
metadata <- read.csv(here('config/3p/metadata.csv')) %>%
  mutate(Kit = factor(Kit, levels = kit_order))
figures <- list()
```

```{r, eval=FALSE}
lapply(objs, function(x) x@project.name)
for (obj in names(objs)){
  Project(objs[[obj]]) <- obj
}
lapply(objs, function(x) x@project.name)
```

# Generate scCoda data

```{r, eval = TRUE}
objs <- lapply(objs, function(obj){
  obj$cell_labels.coarse <- 
    plyr::mapvalues(obj$cell_labels.fine, 
                    from = names(coarse_mapping),
                    to = coarse_mapping)
  obj
})
comp_table <-  lapply(objs, function(x) {
  x@meta.data |> 
    tabyl(orig.ident, cell_labels.coarse) |>
    melt()
  }) |> rbindlist() |>
  dplyr::rename(Sample = 'orig.ident') |>
  # mutate(Sample = case_when(
  #   grepl('FBv4', Sample) ~ gsub('FBv4', 'Fluent_v4', Sample),
  #   grepl('FBv5', Sample) ~ gsub('FBv5', 'Fluent_V', Sample),
  #   grepl('PA_V3', Sample) ~ gsub('PA_V3', 'Parse_v3', Sample),
  #   .default = Sample
  # )) |>
  merge(metadata, by='Sample', all=TRUE) %>%
  select(-c('Sensitivity', 'File'))
comp_table_sccoda <- comp_table |>
  mutate(Kit = as.character(Kit)) |>
  data.table::dcast.data.table(
    Kit + Sample + Individual + Replicate ~ variable, 
    value.var = 'value') %>%
  mutate_all(~replace(., is.na(.), 0)) 
# comp_table <- merge(comp_table, metadata, by='Sample')

write.table(x = comp_table_sccoda, file = here('rds/3p/composition_table_coarse.txt'),
            quote = FALSE, row.names = FALSE, col.names = TRUE, sep = '\t')
```

# Plot

```{r, fig.height=4, warning=FALSE, fig.width=15}
comp_table |> 
  # melt(measure.vars=unique(coarse_mapping)) |> 
  group_by(Sample) |> 
  mutate(percent = 100*value / sum(value)) |>
  ggplot(aes(x = Kit, y = percent,
             group = Kit, color = Kit, shape = paste0(Individual, Replicate))) + 
  geom_point() +
  labs(x = element_blank(), y = 'Proportion of sample',
       color = 'Kit', shape='Sample',
       caption = paste(
         'Portions are of the total capture for a sample.',
         # '* = significant below 0.05 FDR',
         sep='\n')) +
  theme_bw() +
  scale_shape_manual(values = c(1:5))+ 
  theme(axis.text.x = element_blank(), axis.ticks.x = element_blank()) +
  facet_grid(~ factor(variable), scales='free_y')
  ## This is known to be significant from the scCODA results
  # geom_signif(data = data.frame(full_labels = 'pDC',
  #                               label = '*',
  #                               y = .035,
  #                               start = 1.5, end = 3.5,
  #                               outlier = NA),
  #             aes(annotations = label, y_position = y,
  #                 xmin = start, xmax = end),
  #             inherit.aes = FALSE,
  #             tip_length = -0.03,
  #             margin_top = -0.02,
  #             vjust = 2.5,
  #             manual = TRUE, map_signif_level = TRUE)
# pre_vs_post_coarse_comp_boxplot
# ggsave(here('outputs/innate/DC+monocyte/composition_boxplot.png'), height = 4, width = 15)
```

# Save objs

```{r, eval=TRUE}
# saveRDS(objs, here('rds/06_''), compress=FALSE)
```

# Render report

```{r, eval=FALSE}
## This only works if all the variables called in the format file 
## are in the current environment. Check the format file to tweak
## aesthetics 
rmarkdown::render(here('processing_scripts/3p/06-composition.format.Rmd'),
                  output_file = '06-composition.html',
                  output_dir = here('reports/3p'))
```
