==> processing_scripts/3p/08-pseudobulking.runfile.Rmd <==
---
title: 'SC processing'
author: "FHIL\nDerrikGratz"
date: '`r Sys.Date()`'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 

library(tidyverse)  ## General R logic
library(here)       ## Easier specification of file locations
library(yaml)       ## parses config yaml
library(Seurat)
library(reshape2)   ## DF manipulation
# library(clustree)   ## Resolution selection for clustering
# library(ggraph)     ## To render clustree
# library(ggforestplot) ## Striped lines for dotplots
# library(SingleCellExperiment)
library(gridExtra)  ## arrange plots
# library(janitor) ## Tabyl, adorn_rounding
# library(ggrastr) ## Rasterize images for smaller size
library(DT)     ## Datatables
library(DESeq2)  ## Pseudobulk analysis
# library(UCell)  ## Module scoring
library(parallel)
library(data.table)
library(UpSetR)
library(variancePartition)
rbindlist <- data.table::rbindlist
# library(patchwork)
library(gt)
library(emdist)
set.seed(33)
```

# Load data

```{r}
source(here('config/kit_order.R'))
source(here('config/color_palette.R'))
  metadata <- read.csv(here('config/3p/metadata.csv')) %>%
    mutate(Kit = factor(Kit, levels = kit_order_3p))
figures <- list()
```

```{r}
objs <- readRDS(here('rds/3p/07_post_module_scoring.rds'))
genes <- Reduce(intersect, lapply(objs, Features))
objs <- lapply(objs, subset, features = genes)
objs <- lapply(objs, function(x) {
  x$ind <- ifelse(grepl('F1', x$orig.ident), 'F1', 'F5')
  x$group <- paste0(x$cell_labels.fine, '.', x$ind)
  x
})
objs <- lapply(objs, function(x) {
  x$cell_labels.coarse <- case_when(
    x$cell_labels.fine == 'CD4+ T' ~ 'T',
    x$cell_labels.fine == 'CD8+ T' ~ 'T',
    x$cell_labels.fine == 'B naive' ~ 'B',
    x$cell_labels.fine == 'B memory' ~ 'B',
    x$cell_labels.fine == 'Classical monocyte' ~ 'Monocyte',
    x$cell_labels.fine == 'Non-classical monocyte' ~ 'Monocyte',
    .default = x$cell_labels.fine
  )
    x
})
```

```{r}
cellcounts_coarse <- lapply(objs, function(x) {
  table(x$orig.ident, x$cell_labels.coarse) |>
    as.data.table() 
}) |> rbindlist(idcol = 'Kit') |>
  mutate(Individual = gsub('.+_(F.).','\\1',V1)) |>
  group_by(V1) |>
  mutate(prop = N / sum(N))
colnames(cellcounts_coarse) <- c('Kit', 'Sample', 'Celltype', 'Count', 'Individual', 'prop')
cellcounts_coarse_kit <- cellcounts_coarse |>
  group_by(Kit) |> 
    mutate(prop = Count / sum(Count)) |>
  group_by(Kit, Celltype) |> 
  summarize(prop = sum(prop), Count = sum(Count))
write.table(cellcounts_coarse, 
            here('rds/3p/pb_cell_counts.txt'), sep='\t', 
            quote=FALSE, row.names = FALSE)
cellcounts_coarse <- cellcounts_coarse |>
  group_by(Sample) |>
  mutate(prop = Count / sum(Count))
```

# pseudobulking 

```{r}
## subsetting
objs <- lapply(objs, subset, cell_labels.fine %in% c('CD4+ T', 'CD8+ T',
                                                     'B naive', 'B memory',
                                                     'Classical monocyte', 'Non-classical monocyte'))

## Testing downsampling 
objs <- lapply(objs, SetIdent, value='cell_labels.coarse')
objs <- lapply(objs, subset, downsample = 2500)

objs <- lapply(objs, AggregateExpression, 
               group.by = c('orig.ident', 'cell_labels.coarse'),
               assays = 'RNA')
## For scale
objs$Scale$RNA@x <- round(objs$Scale$RNA@x)




pb_obj <- cbind(
  objs$Flex$RNA,
  objs$Fluent_v4$RNA,
  objs$Fluent_V$RNA,
  objs$GEMX3P$RNA,
  objs$NextGEM3P$RNA,
  objs$Parse_v3$RNA,
  objs$Scale$RNA)
# rm(objs)
```

```{r}
label_order <- c(
    # "CD4+ T", "CD8+ T", #'T', "NK",
    # "B naive", "B memory", #'B', "Megakaryocyte",
    # "Classical monocyte", "Non-classical monocyte"#, 'Monocyte',
    # "Dendritic", "pDC",
    # 'Erythrocyte','Granulocyte', 'Unknown'
  'T', 'B', 'Monocyte'
  )
objs <- lapply(objs, function(pb_obj) {
  
  label_order <- c(
    'T', 'B', 'Monocyte'
    # "CD4+ T", "CD8+ T", #'T', "NK",
    # "B naive", "B memory", #'B', "Megakaryocyte",
    # "Classical monocyte", "Non-classical monocyte"#, 'Monocyte',
    # "Dendritic", "pDC",
    # 'Erythrocyte','Granulocyte', 'Unknown'
  )
  pb_obj <- pb_obj$RNA
  coldata <- pb_obj@Dimnames[[2]] |> 
      as.data.table() |> 
      separate(V1, c('Sample', 'celltype'), '_', remove=FALSE) |>
      mutate(Sample = gsub('-', '_', Sample)) |>
      merge(metadata, by = 'Sample') |>
    column_to_rownames('V1') |>
    mutate(celltype = factor(celltype, levels = label_order))
  
  design_formula <- as.formula('~ celltype + Individual')
  design_matrix  <- stats::model.matrix(design_formula, data = coldata)
  coldata <- coldata[colnames(pb_obj),]
  colnames(design_matrix)
  pb_obj <- DESeqDataSetFromMatrix(pb_obj,
                           colData = coldata,
                           design = design_formula,
                           tidy = FALSE)
  assays(pb_obj)$vst <- vst(pb_obj)
  pb_obj
})

```

# Sample dist

```{r}
sampleDists <- 
  lapply(objs, function(pb_obj) {
    dist(t(assay(assays(pb_obj)$vst))) |>
    as.matrix() |>
    as.data.table(keep.rownames = TRUE) |>
    # rownames_to_column('Sample1') |>
    melt() |>
    separate_wider_delim(rn, '_', too_many = 'merge',
                           names = c('Sample1', 'celltype1'), cols_remove = TRUE) |>
      separate_wider_delim(variable, '_', too_many = 'merge',
                           names = c('Sample2', 'celltype2'), cols_remove = TRUE) |>
      separate_wider_regex(Sample1, patterns = c(
        Kit1 = '^.+', '-', Individual1 = '[^-]+$'
      ), cols_remove = TRUE) |>
      separate_wider_regex(Sample2, patterns = c(
        Kit2 = '^.+', '-', Individual2 = '[^-]+$'
      ), cols_remove = TRUE) |>
      mutate(Kit1 = gsub('-', '_', Kit1),
             Kit2 = gsub('-', '_', Kit2)) 
  }) |>
  rbindlist()
```


```{r, fig.width=12}
plotdata <- sampleDists |>
  filter(celltype1 == celltype2 &
           Kit1 == Kit2) |>
  filter(
    (Individual1 == 'F1A' & Individual2 == 'F1B') |
    (Individual1 == 'F5A' & Individual2 == 'F5B')
  ) |>
  mutate(Individual = case_when(
    grepl('F1', Individual1) ~ 'F1',
    grepl('F5', Individual1) ~ 'F5'
  )) |>
  select(-c(Individual1, Individual2, celltype2, Kit2)) |>
  # group_by(Kit1, Individual1, celltype1) |>
  # summarize(y=mean(value)) |>
  rename(Kit1 = 'Kit', celltype1 = 'Celltype') |>
  mutate(Kit = factor(Kit, levels = kit_order_3p),
         Celltype = factor(Celltype, levels = label_order)) |>
  merge(cellcounts_coarse |> group_by(Kit, Celltype, Individual) |> 
          summarize(prop = mean(prop)),
        by=c('Kit', 'Individual', 'Celltype'))
plotdata |>
  ggplot(aes(x=Individual, y=Celltype)) +
  geom_point(aes(size=prop, color=value)) +
  geom_text(aes(label=round(value, 2)), size=2) +
  scale_size_continuous(range=c(5,12)) +
    scale_color_gradient2(low = "#4DAF4A",
                       mid="grey",
                       high = "#FF0000", midpoint = 100, limits=c(0,200), oob=scales::squish) +
  scale_x_discrete(labels = label_function(mode = 'WT')) +
  theme_bw() +
  facet_wrap(~ Kit, nrow = 1) +
  labs(x = 'Individual', y = 'Celltype',
       color = 'Average\nreplicate\ncorrelation', 
       size = 'Proportion\nof cells\nfrom kit')  ->
  figures[['sample_vst_distance_dotplot']]
figures[['sample_vst_distance_dotplot']]
```


```{r}
plotdata <- sampleDists |>
  filter(celltype1 == celltype2,
         Kit1 == Kit2,
         (Individual1 == 'F1A' & Individual2 == 'F1B') |
         (Individual1 == 'F5A' & Individual2 == 'F5B')) |>
  
  mutate(value = value / max(value)) |> 
  # group_by(Kit1, celltype1) |>
  # summarize(y=mean(value)) |>
  rename(Kit1 = 'Kit', celltype1 = 'Celltype', value = 'value') |>
  mutate(Kit = factor(Kit, levels = kit_order_3p),
         Celltype = factor(Celltype, levels = label_order)) |>
  mutate(Sample = paste0(Kit, '_', Individual1)) |>
  merge(cellcounts_coarse, by = c('Kit', 'Sample', 'Celltype')) |>
  mutate(Individual = substr(Individual1, 1, 2)) |>
  mutate(z = value * prop) |>
  group_by(Individual, Kit) |>
  summarize(z = sum(z)) |>
  # group_by(Kit) |>
  # summarize(low = min(z), high = max(z), avg = mean(z)) |>
  janitor::adorn_rounding(digits = 3) |>
  as.data.table() |>
  dcast(Kit ~ Individual, value.var = 'z')
  # mutate(text = paste0(low, ' - ', high))

plotdata |> 
  gt::gt() |>
  # cols_hide(c(low, high, avg)) |>
  # cols_label(Kit = 'Kit', text = 'Sample distance') |>
  data_color(columns = c("F1", 'F5'), method = 'numeric', palette = 'RdYlGn', reverse = TRUE) ->
  
  figures[['pb_vst_dist_table']]
figures[['pb_vst_dist_table']]
```


# PCA

## By kit

```{r}
for (kit in unique(metadata$Kit)) {
  plotdata <- assay(assays(objs[[kit]])$vst)
  plotdata <- plotdata[,grepl(kit, gsub('-', '_', colnames(plotdata)))]
  plotdata <- stats::prcomp(t(plotdata))$x |>
    as.data.frame() |>
    rownames_to_column('sample') |>
    separate(sample, c('Sample', 'celltype'), '_', remove = TRUE) |>
    mutate(Sample = gsub('-', '_', Sample)) |>
    merge(metadata, by = 'Sample') 
  ggplot(plotdata, aes(x=PC1, y=PC2)) +
  # geom_point(aes(shape = Kit, color = celltype)) + 
  geom_point(aes(shape = paste0(Individual, Replicate), color = celltype)) + 
  # stat_ellipse(aes(group = celltype, color = celltype, type = "norm", level = 0.67)) +
  # stat_ellipse(aes(color = Kit, type = "norm", level = 0.67), show.legend = FALSE) +
  scale_color_manual(values=color_palette$cell_colors) +
  theme_bw() +
  labs(x='PC1', y='PC2', color='Celltype', shape='Sample', title=kit) ->
    figures[[paste0('PCA_', kit)]]
}
figures[[paste0('PCA_', kit)]]
figures[['PCA_Flex']]
```

<!-- ## Euclidean distance PCA -->

<!-- ```{r} -->
<!-- plotdata <- -->
<!--   pca(assay(assays(pb_obj)$vst), nfeatures = 200, npcs = 20) |> -->
<!--     dist() |> -->
<!--     as.matrix() |> -->
<!--     as.data.frame() |> -->
<!--     rownames_to_column('sample1') |> -->
<!--     melt(variable.name = 'sample2') |> -->
<!--     separate(sample1, c('sample1', 'celltype1'), '_', remove = TRUE) |> -->
<!--     separate(sample2, c('sample2', 'celltype2'), '_', remove = TRUE) |> -->
<!--   separate_wider_regex(sample1, c(kit1 = '.+', '-', sample1='.+$')) |> -->
<!--   separate_wider_regex(sample2, c(kit2 = '.+', '-', sample2='.+$')) |> -->
<!--     filter(((sample1 == 'F1A' & sample2 == 'F1B') | (sample1 == 'F5A' & sample2 == 'F5B')) & -->
<!--            celltype1 == celltype2 & -->
<!--              kit1 == kit2) |> -->
<!--     mutate(ind = ifelse(grepl('F1', sample2), 'F1', 'F5')) |> -->
<!--   arrange(celltype1) |> -->
<!--   select(c(kit1, celltype1, sample1, sample2, value, ind)) |> -->
<!--   mutate(kit1 = factor(gsub('-', '_', kit1), levels = kit_order_3p), -->
<!--          celltype1 = factor(celltype1, levels = label_order)) |> -->
<!--   merge(cellcounts_coarse_kit, by.x = c('kit1', 'celltype1'), by.y=c('Kit', 'Celltype')) -->
<!-- # } -->
<!-- ggplot(plotdata , -->
<!--   aes(x=ind, y=celltype1)) + -->
<!--   geom_point(aes(size=prop, color=value)) + -->
<!--   geom_text(aes(label=round(value, 2)), size=2) + -->
<!--   scale_size_continuous(range=c(5,12)) + -->
<!--   scale_size_continuous(range=c(5,12)) + -->
<!--     scale_color_gradient2(low = "#4DAF4A", -->
<!--                        mid="#FFFFCC", -->
<!--                        high = "#FF0000", -->
<!--                        midpoint=max(plotdata$value)/2, limits = c(0, max(plotdata$value))) + -->
<!--     labs(x = 'Sample', y = 'Celltype',color = 'Average\nreplicate\ncorrelation', -->
<!--        size = 'Proportion\nof cells\nfrom kit') + -->
<!--     facet_wrap(~ kit1, ncol=7, nrow = 1, labeller = labeller(kit1 = label_function())) -> -->
<!--     figures[['pca_dist_euc_pairs']] -->

<!-- plotdata |> -->
<!--   group_by(kit1, celltype1) |> -->
<!--   summarize(value = mean(value), prop = mean(prop)) |> -->
<!-- ggplot( -->
<!--          aes(x=kit1, -->
<!--              y=celltype1, -->
<!--              size=prop)) + -->
<!--   geom_point(aes(size=prop, color=value)) + -->
<!--   geom_text(aes(label=round(value, 2)), size=2) + -->
<!--   scale_size_continuous(range=c(5,12)) + -->
<!--   scale_size_continuous(range=c(5,12)) + -->
<!--     scale_color_gradient2(low = "#4DAF4A", -->
<!--                        mid="#FFFFCC", -->
<!--                        high = "#FF0000", -->
<!--                        midpoint=max(plotdata$value)/2, limits = c(0, max(plotdata$value))) + -->
<!--     labs(x = 'Sample', y = 'Celltype',color = 'Average\nreplicate\ncorrelation', -->
<!--        size = 'Proportion\nof cells\nfrom kit') -> -->
<!--     figures[['pca_dist_euc_mean']] -->
<!-- ``` -->

<!-- ```{r, fig.width=11, fig.height=5} -->
<!-- figures[['pca_dist_euc_mean']] -->
<!-- ``` -->

<!-- ```{r, fig.width=14, fig.height=5} -->
<!-- figures[['pca_dist_euc_pairs']] -->
<!-- ``` -->


<!-- ```{r, eval=FALSE} -->
<!-- euc_data <- expand.grid(unique(metadata$Kit), unique(metadata$Individual), label_order) |> -->
<!--   rename(Var1 = 'Kit', Var2 = 'Individual', Var3 = 'Celltype') -->
<!-- for (kit in unique(metadata$Kit)) { -->
<!--   plotdata <- assay(vsd) -->
<!--   plotdata <- plotdata[,grepl(kit, gsub('-', '_', colnames(plotdata)))] -->
<!--   plotdata <- pca(plotdata, npcs = 40) |> -->
<!--     as.data.frame() |> -->
<!--     rownames_to_column('sample') |> -->
<!--     separate(sample, c('Sample', 'celltype'), '_', remove = TRUE) |> -->
<!--     mutate(Sample = gsub('-', '_', Sample)) |> -->
<!--     merge(metadata, by = 'Sample') -->
<!--   for (ind in c('F1', 'F5')) { -->
<!--     for (cell in unique(plotdata$celltype)) { -->
<!--      tmp <- plotdata |> -->
<!--        filter(Individual == ind, celltype == cell) |> -->
<!--        select(starts_with('PC')) |> -->
<!--        dist(method='euclidean') -->
<!--     } -->
<!--   } -->

<!-- } -->
<!-- ``` -->

# Corr scatterplot

```{r}
corr_scatterplot <- function() {
  plotdata <- assays(pb_obj)$vst |> 
    assay() |> 
    melt() |> 
    dplyr::rename(Gene = Var1) |>
    separate(Var2, into = c('Sample', 'Celltype'), sep='_') |>
    mutate(Sample = gsub('-', '_', Sample, fixed=TRUE)) |>
    merge(metadata, by='Sample')
  plotdata <- merge(plotdata, plotdata, by = c('Kit', 'Gene'))
  ggplot(plotdata, aes(x=))
}
```


# Corr mat

## funcs

spearman and pearson are the same with VST data. Already scaled? 

```{r}
obj_cor <- function(obj, kit, cell_type, ngenes = 200,
                    method = 'pearson', 
                    features_select = c('celltype', 'kit+celltype'),
                    features = NULL) {
  obj <- subset(obj, select=colData(obj)$celltype==cell_type)
  if (features_select == 'celltype') {rv <- rowSums(assay(obj))}
  obj <- subset(obj, select=colData(obj)$Kit==kit)
  if (features_select == 'kit+celltype') {rv <- rowSums(assay(obj))}
  if (is.null(features)) {
    features <- names(head(sort(rv, decreasing=TRUE), ngenes))
  }
  # if ()
  obj <- subset(obj, rownames(obj) %in% features)
  obj <- assay(obj) |>
    cor(method = method) |>
    as.data.table(keep.rownames = TRUE) |>
    melt(id.var='rn') |> 
    separate_wider_delim(rn, '_', too_many = 'merge',
                         names = c('Sample1', 'celltype1'), cols_remove = TRUE) |>
    separate_wider_delim(variable, '_', too_many = 'merge',
                         names = c('Sample2', 'celltype2'), cols_remove = TRUE) |>
    separate_wider_regex(Sample1, patterns = c(
      Kit1 = '^.+', '-', Individual1 = '[^-]+$'
    ), cols_remove = TRUE) |>
    separate_wider_regex(Sample2, patterns = c(
      Kit2 = '^.+', '-', Individual2 = '[^-]+$'
    ), cols_remove = TRUE) |>
    mutate(Kit1 = gsub('-', '_', Kit1),
           Kit2 = gsub('-', '_', Kit2))
  return(obj)
}
```

```{r}
corr_dotplot <- function(corrdata, midpoint = 0.5) {
  corrdata  |>
  merge(cellcounts_coarse_kit, by=c('Kit', 'Celltype')) |>
  ggplot(aes(x=Kit, y=Celltype)) +
  geom_point(aes(size=prop, color=y)) +
  geom_text(aes(label=round(y, 4)), size=2) +
  scale_size_continuous(range=c(5,12)) +
  scale_color_gradient2(high = "#4DAF4A",
                       mid="#FFFFCC",
                       low = "#FF0000", 
                     midpoint=midpoint) +
  scale_x_discrete(labels = label_function()) +
  theme_bw() +
  labs(x = 'Kit', y = 'Celltype',
       color = 'Average\nreplicate\ncorrelation', 
       size = 'Proportion\nof cells\nfrom kit')
}
corr_dotplot_byind <- function(corrdata, midpoint = 0.5) {
  corrdata  |>
  merge(cellcounts_coarse_kit, by=c('Kit', 'Celltype')) |>
  ggplot(aes(x=Individual1, y=Celltype)) +
  geom_point(aes(size=prop, color=y)) +
  geom_text(aes(label=round(y, 4)), size=2) +
  scale_size_continuous(range=c(5,12)) +
  scale_color_gradient2(high = "#4DAF4A",
                       mid="#FFFFCC",
                       low = "#FF0000", 
                     midpoint=midpoint) +
  scale_x_discrete(labels = label_function()) +
  theme_bw() +
  labs(x = 'Sample', y = 'Celltype',
       color = 'Average\nreplicate\ncorrelation', 
       size = 'Proportion\nof cells\nfrom kit') +
  facet_wrap(~ Kit, ncol=7, nrow = 1, labeller = labeller(kit1 = label_function())) 
}
```

## Invidiual variance genes

```{r}
features <- results(pb_obj, name = 'Individual_F5_vs_F1') |> as.data.frame() |> arrange(padj) |> head(n=500) |> rownames()
for (kit in unique(metadata$Kit)) {
  plotdata <- tibble()
  for (celltype in unique(colData(objs[[kit]])$celltype)) {
    if (nrow(filter(cellcounts_coarse_kit, Kit == kit, Celltype == celltype)) > 0){
      # # var_genes <- adsf |>
      #   as.data.frame() |> 
      #   arrange(padj) |> 
      #   head(n=200) |> 
      #   rownames()
      plotdata1 <- obj_cor(assays(objs[[kit]])$vst, kit, celltype, ngenes=200, method='pearson', features=features, features_select = 'kit+celltype') |>
      filter(celltype1 == celltype2) |>
      # filter((Individual1 == 'F1A' & Individual2 == 'F1B') |
      #      (Individual1 == 'F5A' & Individual2 == 'F5B')) |>
      mutate(celltype1 = factor(celltype1, levels=label_order),
             celltype2 = factor(celltype2, levels=label_order)) |>
      arrange(celltype1, celltype2) |>
      distinct()
      plotdata <- rbind(plotdata, plotdata1)
    }
  }
  corr_data_bykit <- rbind(corr_data_bykit, plotdata)
  
  
  ggplot(distinct(plotdata),
         # aes(x=paste0(Sample1, '_', celltype1),
         #     y=paste0(Sample2, '_', celltype2), 
         aes(x=Individual1,
             y=Individual2,
             label=round(value,2),
             fill=value)) +
    geom_tile() +
    geom_text() +
    scale_fill_gradient2(high = "#4DAF4A",
                       mid="#FFFFCC",
                       low = "#FF0000", 
                       midpoint=0.95) +
    labs(x = 'Sample 1', y = 'Sample 2', fill = 'Corr', title = kit) +
    facet_wrap(~ celltype1, drop = FALSE, ncol=6, nrow = 4) ->
    figures[[paste0('corr_mat_pearson', kit)]]
  print(figures[[paste0('corr_mat_pearson', kit)]])
}
```

## Pathway genes

```{r}
# library(msigdbr)
pathways <- msigdbr(species = "human", category = "C2") |>
 filter(gs_subcat  == 'CP:BIOCARTA')
pathways <- pathways %>% dplyr::select(gene_symbol, gs_name) %>% unstack()
features <- unique(unname(unlist(pathways)))
for (kit in unique(metadata$Kit)) {
  plotdata <- tibble()
  for (celltype in unique(colData(objs[[kit]])$celltype)) {
    if (nrow(filter(cellcounts_coarse_kit, Kit == kit, Celltype == celltype)) > 0){
      # # var_genes <- adsf |>
      #   as.data.frame() |> 
      #   arrange(padj) |> 
      #   head(n=200) |> 
      #   rownames()
      plotdata1 <- obj_cor(assays(objs[[kit]])$vst, kit, celltype, ngenes=2000, method='pearson', features=features, features_select = 'kit+celltype') |>
      filter(celltype1 == celltype2) |>
      # filter((Individual1 == 'F1A' & Individual2 == 'F1B') |
      #      (Individual1 == 'F5A' & Individual2 == 'F5B')) |>
      mutate(celltype1 = factor(celltype1, levels=label_order),
             celltype2 = factor(celltype2, levels=label_order)) |>
      arrange(celltype1, celltype2) |>
      distinct()
      plotdata <- rbind(plotdata, plotdata1)
    }
  }
  corr_data_bykit <- rbind(corr_data_bykit, plotdata)
  
  
  ggplot(distinct(plotdata),
         # aes(x=paste0(Sample1, '_', celltype1),
         #     y=paste0(Sample2, '_', celltype2), 
         aes(x=Individual1,
             y=Individual2,
             label=round(value,2),
             fill=value)) +
    geom_tile() +
    geom_text() +
    scale_fill_gradient2(high = "#4DAF4A",
                       mid="#FFFFCC",
                       low = "#FF0000", 
                       midpoint=0.95) +
    labs(x = 'Sample 1', y = 'Sample 2', fill = 'Corr', title = kit) +
    facet_wrap(~ celltype1, drop = FALSE, ncol=6, nrow = 4) ->
    figures[[paste0('corr_mat_pearson', kit)]]
  print(figures[[paste0('corr_mat_pearson', kit)]])
}
```


## across kits

```{r}
corr_data_all <- list()
for (kit in unique(metadata$Kit)) {
  for (celltype in unique(colData(objs[[kit]])$celltype)) {
    if (nrow(filter(cellcounts_coarse_kit, Kit == kit, Celltype == celltype)) > 0){
        plotdata <- obj_cor(assays(objs[[kit]])$vst, kit, celltype, ngenes=500,
                            method='spearman',
                            features_select = 'celltype'
                            ) |>
      filter((Individual1 == 'F1A' & Individual2 == 'F1B') |
           (Individual1 == 'F5A' & Individual2 == 'F5B')) |>
      filter(celltype1 == celltype2) |>
      mutate(celltype1 = factor(celltype1, levels=label_order),
             celltype2 = factor(celltype2, levels=label_order)) |>
      arrange(celltype1, celltype2) |>
        distinct()
      corr_data_all <- rbind(corr_data_all, plotdata)
    }
  }
}
corr_data_all <- corr_data_all |>
  mutate(#Sample = paste0(Kit1, '_', Individual1),
         Kit = factor(Kit1, levels = kit_order_3p),
         Celltype = factor(celltype1, levels = label_order)) |>
  select(-c(celltype1, Kit1)) |>
  filter(!is.na(celltype)) |>
  group_by(Kit, Individual1, Celltype) |>
  summarize(y = mean(value))
```


```{r, fig.height=5, fig.width=12}
corr_data_all |>
  mutate(y=round(y,2)) |>
  corr_dotplot_byind(midpoint = 0.85) ->
  figures[['rep_corr_summary_all']]
figures[['rep_corr_summary_all']]
```

### 200

```{r, eval=TRUE, fig.width=12, fig.height=4}
corr_data_bykit <- tibble()
for (kit in unique(metadata$Kit)) {
  plotdata <- tibble()
  for (celltype in unique(colData(objs[[kit]])$celltype)) {
    if (nrow(filter(cellcounts_coarse_kit, Kit == kit, Celltype == celltype)) > 0){
      # # var_genes <- adsf |>
      #   as.data.frame() |> 
      #   arrange(padj) |> 
      #   head(n=200) |> 
      #   rownames()
      plotdata1 <- obj_cor(assays(objs[[kit]])$vst, kit, celltype, ngenes=200, method='pearson', features_select = 'kit+celltype') |>
      filter(celltype1 == celltype2) |>
      # filter((Individual1 == 'F1A' & Individual2 == 'F1B') |
      #      (Individual1 == 'F5A' & Individual2 == 'F5B')) |>
      mutate(celltype1 = factor(celltype1, levels=label_order),
             celltype2 = factor(celltype2, levels=label_order)) |>
      arrange(celltype1, celltype2) |>
      distinct()
      plotdata <- rbind(plotdata, plotdata1)
    }
  }
  corr_data_bykit <- rbind(corr_data_bykit, plotdata)
  
  
  ggplot(distinct(plotdata),
         # aes(x=paste0(Sample1, '_', celltype1),
         #     y=paste0(Sample2, '_', celltype2), 
         aes(x=Individual1,
             y=Individual2,
             label=round(value,2),
             fill=value)) +
    geom_tile() +
    geom_text() +
    scale_fill_gradient2(high = "#4DAF4A",
                       mid="#FFFFCC",
                       low = "#FF0000", 
                       midpoint=0.9) +
    labs(x = 'Sample 1', y = 'Sample 2', fill = 'Corr', title = kit) +
    facet_wrap(~ celltype1, drop = FALSE, ncol=6, nrow = 4) ->
    figures[[paste0('corr_mat_pearson', kit)]]
  print(figures[[paste0('corr_mat_pearson', kit)]])
}

```

```{r, fig.width=12}
corr_data_bykit |>
  mutate(Kit = factor(Kit1, levels = kit_order_3p),
         Celltype = factor(celltype1, levels = label_order)) |>
  select(-c(celltype1, Kit1)) |>
  filter(!is.na(celltype))  |>
  group_by(Kit, Individual1, Celltype) |>
  summarize(y=mean(value)) |> 
  corr_dotplot_byind(midpoint = 0.9) ->
  figures[['rep_corr_summary_bykit']]
figures[['rep_corr_summary_bykit']]
```

### 2000

```{r, eval=TRUE, fig.width=12, fig.height=4}
corr_data_bykit <- tibble()
for (kit in unique(metadata$Kit)) {
  plotdata <- tibble()
  for (celltype in unique(colData(objs[[kit]])$celltype)) {
    if (nrow(filter(cellcounts_coarse_kit, Kit == kit, Celltype == celltype)) > 0){
      # # var_genes <- adsf |>
      #   as.data.frame() |> 
      #   arrange(padj) |> 
      #   head(n=200) |> 
      #   rownames()
      plotdata1 <- obj_cor(assays(objs[[kit]])$vst, kit, celltype, ngenes=2000, method='spearman', features_select = 'kit+celltype') |>
      filter(celltype1 == celltype2) |>
      # filter((Individual1 == 'F1A' & Individual2 == 'F1B') |
      #      (Individual1 == 'F5A' & Individual2 == 'F5B')) |>
      mutate(celltype1 = factor(celltype1, levels=label_order),
             celltype2 = factor(celltype2, levels=label_order)) |>
      arrange(celltype1, celltype2) |>
      distinct()
      plotdata <- rbind(plotdata, plotdata1)
    }
  }
  corr_data_bykit <- rbind(corr_data_bykit, plotdata)
  
  
  ggplot(distinct(plotdata),
         # aes(x=paste0(Sample1, '_', celltype1),
         #     y=paste0(Sample2, '_', celltype2), 
         aes(x=Individual1,
             y=Individual2,
             label=round(value,2),
             fill=value)) +
    geom_tile() +
    geom_text() +
    scale_fill_gradient2(high = "#4DAF4A",
                       mid="#FFFFCC",
                       low = "#FF0000", 
                       midpoint=0.9) +
    labs(x = 'Sample 1', y = 'Sample 2', fill = 'Corr', title = kit) +
    facet_wrap(~ celltype1, drop = FALSE, ncol=6, nrow = 4) ->
    figures[[paste0('corr_mat_pearson', kit)]]
  print(figures[[paste0('corr_mat_pearson', kit)]])
}

```

```{r}
corr_data_bykit |>
  mutate(Kit = factor(Kit1, levels = kit_order_3p),
         Celltype = factor(celltype1, levels = label_order)) |>
  select(-c(celltype1, Kit1)) |>
  filter(!is.na(celltype))  |>
  group_by(Kit, Celltype) |>
  summarize(y=mean(value)) |> 
  corr_dotplot(midpoint=-.9) ->
  figures[['rep_corr_summary_bykit']]
figures[['rep_corr_summary_bykit']]
```

# Variance partitioning

```{r}

```


```{r, eval=FALSE}
run_varpart <- function(obj,
                        form = ~ (1 | celltype) + (1 | Individual) + (1 | Replicate),
                        BPPARAM = BiocParallel::MulticoreParam(),
                        ngenes = 2000) {
  # obj  <- obj[ , obj$Kit == kit]
  isexpr <- rowSums(fpm(obj) > 1) >= 0.5 * ncol(obj)
  
  # compute log2 Fragments Per Million
  # Alternatively, fpkm(), vst() or rlog() could be used
  quantLog <- log2(fpm(obj)[isexpr, ] + 1)
  if (ngenes < nrow(quantLog)) {
    quantLog <- quantLog[1:ngenes,]
  }
  varPart <- fitExtractVarPartModel(quantLog,
                                    form, 
                                    as.data.frame(obj@colData),
                                    BPPARAM = BPPARAM)
  varPart
}

# form <- ~ (1 | Kit) + (1 | celltype) + (1 | Kit:celltype) + (1 | Individual) + (1 | Replicate) + (1 | Kit:Replicate) + (1 | Kit:Individual)
# form <- ~ (1 | Kit) + (1 | Kit:Individual) + (1 | Kit:Replicate) + (1 | Kit:celltype)
# form <- ~ (1 | Kit) + (1 | celltype) + (Kit + 0 | Individual) + (Kit + 0 | Replicate) + (Kit + 0 | celltype) 
 #+ (1 | Individual:Replicate)
# form <- ~ (1 | Individual) + (1 | Replicate) + (1 | Individual:Replicate)

# varPart <- run_varpart(objs$Flex, ngenes = 5000, form = ~ (1 | celltype) + (1 | Individual) + (1 | Replicate), BPPARAM = BiocParallel::SerialParam())
# Run variancePartition analysis
varPart <- lapply(objs, run_varpart, BPPARAM = BiocParallel::SerialParam(), ngenes = 50000, form = ~ (1 | celltype) + (1 | Individual) + (1 | Replicate))
saveRDS(varPart, here('rds/3p/varpart_results.Rds'))
```

```{r}
varPart <- readRDS(here('rds/3p/varpart_results.Rds'))
```


```{r, fig.width = 18, fig.height=6}
plotdata <- lapply(varPart, function(x) {
  as.data.frame(x) |> 
    as.data.table() |> 
    melt()
  }) |>
  rbindlist(idcol = 'Kit') |>
  mutate(Kit = factor(Kit, kit_order_3p), 
         variable = factor(variable, c('celltype', 'Individual', 'Replicate', 'Residuals')))
colnames(plotdata) <- c('Kit', 'Category', 'Value')
ggplot(plotdata, aes(x=Kit, y=Value, fill = Kit))  +
  # geom_violin() +
  geom_boxplot(outliers = FALSE) +
  scale_y_continuous(labels = scales::percent) +
  scale_x_discrete(labels = label_function()) +
  theme(axis.text.x = element_text(angle=45, hjust=1)) +
  scale_fill_manual(values = color_palette$kits, labels = label_function()) + 
  facet_wrap(~ Category, scales='free_y', nrow=1) +
  labs(x='Kit', y='Percent of variance in expression explained')
# vp <- sortCols(varPart)
# 
# # Figure 1a
# # Bar plot of variance fractions for the first 10 genes
# plotPercentBars(vp[1:10, ])
# variancePartition::plotVarPart(varPart)
```

## Replicability metric

```{r}
plotdata <- lapply(varPart, function(x) {
  as.data.frame(x) |> 
    mutate(r = celltype + Individual / rowSums(across(where(is.numeric)))) |>
    select(r)
}) |>
  rbindlist(idcol = 'Kit') |>
  mutate(Kit = factor(Kit, kit_order_3p))

ggplot(plotdata, aes(x=Kit, y=r, fill = Kit))  +
  # geom_violin() +
  geom_boxplot(outliers = FALSE) +
  scale_y_continuous(labels = scales::percent) +
  scale_x_discrete(labels = label_function()) +
  theme(axis.text.x = element_text(angle=45, hjust=1)) +
  scale_fill_manual(values = color_palette$kits, labels = label_function()) + 
  labs(x='Kit', y='Replicability (per gene)')
```

# DESeq

```{r, eval = TRUE}
# pb_obj <- DESeq(pb_obj)
objs <- lapply(objs, function(obj) {
  design(obj) <- ~ 0 + celltype + Individual
  obj
})  
objs <- lapply(objs, DESeq, full = ~ 0 + celltype + Individual)
saveRDS(objs, here('rds/3p/pseudobulk_objs_split_by_kit_allgenes_downsamplecells_coarse_celltypes.Rds'), compress = FALSE)
```

```{r,eval=TRUE}
objs <- readRDS(here('rds/3p/pseudobulk_objs_split_by_kit_allgenes_downsamplecells_coarse_celltypes.Rds'))
```

# Render report

```{r, eval=TRUE}
## This only works if all the variables called in the format file 
## are in the current environment. Check the format file to tweak
## aesthetics 
rmarkdown::render(here('processing_scripts/3p/08-pseudobulking.format.Rmd'),
                  output_file = '08-pseudobulking.html',
                  output_dir = here('reports/3p/'))
```
