==> processing_scripts/3p/09-dge.format.Rmd <==
---
title: 'BM 3P WT DGE'
author: |
  | FHIL
  | Derrik Gratz
date: '`r Sys.Date()`'
output: 
  rmdformats::robobook:
    fig_width: 9
    toc_depth: 1
    css: '../../config/robobook.css'
---

```{r, include=FALSE}
library(knitr)
knitr::opts_chunk$set(echo = FALSE, warning = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 
```

# DGE {.tabset}

Since we don't have any experimental conditions, I figured it would be easiest to to DGE on celltypes. I looked at 4 contrasts. We don't have to use all or any of these, but I just wanted to explore.

1. B vs. T
2. T vs. Monocyte
3. Monocyte vs. B
4. B vs. T and Monocyte
5. T vs. B and Monocyte
6. Monocyte vs. T and B

I show the overlap significantly upregulated genes in a few comparisons. The counts are high and not very overlapping.

```{r dge_overlap, eval=TRUE, results='asis'}
#Then a breakdown of all kits (this may be excessive)
out <- NULL
for (contrast in contrast_matrix_formula$contrast_names) {
  # chunktext <- '\n## {{contrast}} {.tabset .unlist}\n'
  # out = c(out, knit_expand(text = chunktext))
  # for (kit in unique(metadata$Kit)) {
  # \n\n::::{class='fluid-row'}\n\n:::{class='col-md-4'}
  # \n\n::::{class='fluid-row'}\n\n:::{class='col-md-8'}
  # \n\n:::\n\n::::\n\n***\n
    chunktext <- "
\n## {{contrast}} {.unlist}
\n```{r}
figures[[paste0('euler_', '{{contrast}}')]]
figures[[paste0('upset_', '{{contrast}}')]]
\n```
\n***\n
"
    out = c(out, knit_expand(text = chunktext))
  # }
}

cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```

# DGE volcanos {.tabset}

```{r dge_volcanos, eval=TRUE, results='asis'}
#Then a breakdown of all kits (this may be excessive)
out <- NULL
for (contrast in contrast_matrix_formula$contrast_names) {
  chunktext <- '\n## {{contrast}} {.tabset .unlist}\n'
  out = c(out, knit_expand(text = chunktext))
  for (kit in unique(metadata$Kit)) {
    chunktext <- '
    \n### {{kit}} {.unlist}\n
    \n```{r, fig.width=8, fig.height=5}
    \nfigures[["dge_volcano"]][[paste0("{{kit}}",".","{{contrast}}")]]
    \n```\n
    '
    out = c(out, knit_expand(text = chunktext))
  }
}

cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```

# DGE correlation {.tabset}

```{r dge_corplot, eval=TRUE, results='asis'}
#Then a breakdown of all kits (this may be excessive)
out <- NULL
for (contrast in contrast_matrix_formula$contrast_names) {
  chunktext <- "
  \n## {{contrast}} {.unlist}\n
  \n```{r, fig.width=10, fig.height=10}
  \nfigures[[paste0('corr_pairplot_', '{{contrast}}')]]
  \n```\n
  "
  out = c(out, knit_expand(text = chunktext))
}

cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```


# GSEA

We then turn to GSEA to see if we can make sense of the high numbers of DGE in the comparisons. We can look both at the total set of pathways tested to see if there are any strange positive results, and we can focus on expected positive results. A volcano plot may not be the best way to represent this data, as the two axes are not fully independent, but it works for now to highlight the point. **Note that while this says log fold change, it actually means Normalized Enrichment Score**

## All pathways {.tabset}

```{r, eval=TRUE, results='asis'}
#Then a breakdown of all kits (this may be excessive)
out <- NULL
for (contrast in contrast_matrix_formula$contrast_names) {
  chunktext <- '\n### {{contrast}} {.tabset .unlist}\n\n'
  out = c(out, knit_expand(text = chunktext))
  for (kit in unique(metadata$Kit)) {
    chunktext <- '
    \n#### {{kit}} {.unlist}\n
    \n```{r, fig.width=8, fig.height=5}
    figures[["gsea_volcano"]][[paste0("{{kit}}",".","{{contrast}}")]]
    \n```\n
    '
    out = c(out, knit_expand(text = chunktext))
  }
}

cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```

## Pathways of interest

```{r, fig.width=15, fig.height=10}
figures[['gsea_dotplot']]
```
