==> processing_scripts/3p/09-dge.runfile.Rmd <==
---
title: 'SC processing'
author: "FHIL\nDerrikGratz"
date: '`r Sys.Date()`'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 

library(tidyverse)  ## General R logic
library(here)       ## Easier specification of file locations
library(yaml)       ## parses config yaml
library(reshape2)   ## DF manipulation
# library(clustree)   ## Resolution selection for clustering
# library(ggraph)     ## To render clustree
library(ggforestplot) ## Striped lines for dotplots
# library(SingleCellExperiment)
library(gridExtra)  ## arrange plots
# library(janitor) ## Tabyl, adorn_rounding
# library(ggrastr) ## Rasterize images for smaller size
library(DT)     ## Datatables
library(DESeq2)  ## Pseudobulk analysis
# library(UCell)  ## Module scoring
# library(parallel)
library(BiocParallel)
library(data.table)
library(UpSetR)
rbindlist <- data.table::rbindlist
# library(patchwork)
library(gt)
library(fgsea)
library(msigdbr)
set.seed(33)
```

# Load data

```{r}
source(here('config/kit_order.R'))
source(here('config/color_palette.R'))
  metadata <- read.csv(here('config/3p/metadata.csv')) %>%
    mutate(Kit = factor(Kit, levels = kit_order_3p))
figures <- list()
```


```{r}
objs <- readRDS(here('rds/3p/pseudobulk_objs_split_by_kit_allgenes_downsamplecells_coarse_celltypes.Rds'))
```


# Define contrast matrix

```{r}
design_matrix  <- stats::model.matrix(objs$Flex@design, data = objs$Flex@colData)
colnames(design_matrix) |> make.names()
```
```{r}
# I make these very explicit to track all terms. They get simplified out at the end. See the output of this chunk
contrast_matrix_formula <- data.frame(
  contrast_names = c(
    'B.vs.T',
    'T.vs.M',
    'M.vs.B',
    # 'CD8.vs.CD4',
    'B.vs.all',
    'T.vs.all',
    'M.vs.all'
  ),
  numerators = c(
    'celltypeB',
    'celltypeT',
    'celltypeMonocyte',
    'celltypeB',
    'celltypeT',
    'celltypeMonocyte'
  ),
  denominators = c(
    'celltypeT',
    'celltypeMonocyte',
    'celltypeB',
    # 'celltypeCD4..T',
    '.5 * ( celltypeT + celltypeMonocyte )',
    '.5 * ( celltypeB + celltypeMonocyte )',
    '.5 * ( celltypeT + celltypeB )'
  )
)
contrast_matrix_expanded <- contrast_matrix_formula
```

```{r, eval=FALSE}
# I make these very explicit to track all terms. They get simplified out at the end. See the output of this chunk
contrast_matrix_formula <- data.frame(
  contrast_names = c(
    'B.vs.T',
    'T.vs.M',
    'M.vs.B',
    # 'CD8.vs.CD4',
    'B.vs.all',
    'T.vs.all',
    'M.vs.all'
  ),
  numerators = c(
    '0.5 * ( celltypeB.memory + celltypeB.naive )',
    '0.5 * ( celltypeCD4..T + celltypeCD8..T )',
    '0.5 * ( celltypeNon.classical.monocyte + celltypeClassical.monocyte )',
    # 'celltypeCD8..T',
    '0.5 * ( celltypeB.memory + celltypeB.naive )',
    '0.5 * ( celltypeCD4..T + celltypeCD8..T )',
    '0.5 * ( celltypeNon.classical.monocyte + celltypeClassical.monocyte )'
  ),
  denominators = c(
    '0.5 * ( celltypeCD4..T + celltypeCD8..T )',
    '0.5 * ( celltypeNon.classical.monocyte + celltypeClassical.monocyte )',
    '0.5 * ( celltypeB.memory + celltypeB.naive )',
    # 'celltypeCD4..T',
    '.25 * ( celltypeCD4..T + celltypeCD8..T + 
      celltypeNon.classical.monocyte + celltypeClassical.monocyte )',
    '.25 * ( celltypeB.memory + celltypeB.naive + 
      celltypeNon.classical.monocyte + celltypeClassical.monocyte )',
    '.25 * ( celltypeCD4..T + celltypeCD8..T + 
      celltypeB.memory + celltypeB.naive )'
  )
)
contrast_matrix_expanded <- contrast_matrix_formula
```

```{r contrasts_with_kit, eval=FALSE}
# I make these very explicit to track all terms. They get simplified out at the end. See the output of this chunk
contrast_matrix_formula <- data.frame(
  contrast_names = c(
    'B.vs.T',
    'T.vs.M',
    'M.vs.B',
    'CD8.vs.CD4',
    # 'CD4.F5.vs.F1',
    # 'CD4.vs.otherCD4',
    'B.vs.A'
  ),
  numerators = c(
    'X.Intercept. + Kit{{KIT}} + 0.5 * ( celltypeB.memory + celltypeB.naive + Kit{{KIT}}.celltypeB.memory + Kit{{KIT}}.celltypeB.naive )',
    'X.Intercept. + Kit{{KIT}} + 0.5 * ( celltypeCD8..T + Kit{{KIT}}.celltypeCD8..T )',
    'X.Intercept. + Kit{{KIT}} + 0.5 * ( celltypeNon.classical.monocyte + celltypeClassical.monocyte + Kit{{KIT}}.celltypeNon.classical.monocyte + Kit{{KIT}}.celltypeClassical.monocyte )',
    'X.Intercept. + Kit{{KIT}} + celltypeCD8..T   + Kit{{KIT}}.celltypeCD8..T',
    # 'X.Intercept. + Kit{{KIT}} + (1/6) * Kit{{KIT}}', # last term to cancel with denom
    'X.Intercept. + Kit{{KIT}} + Kit{{KIT}}.ReplicateB'
  ),
  denominators = c(
    'X.Intercept. + Kit{{KIT}} + 0.5 * ( celltypeCD8..T + Kit{{KIT}}.celltypeCD8..T )',
    'X.Intercept. + Kit{{KIT}} + 0.5 * ( celltypeNon.classical.monocyte + celltypeClassical.monocyte + Kit{{KIT}}.celltypeNon.classical.monocyte + Kit{{KIT}}.celltypeClassical.monocyte )',
    'X.Intercept. + Kit{{KIT}} + 0.5 * ( celltypeB.memory + celltypeB.naive + Kit{{KIT}}.celltypeB.memory + Kit{{KIT}}.celltypeB.naive )',
    'X.Intercept. + Kit{{KIT}}',
    # 'X.Intercept. + (1/6) * ( KitNextGEM3P + KitGEMX3P + KitFluent_V + KitFluent_v4 + KitScale + KitParse_v3 )',
    'X.Intercept. + Kit{{KIT}}'
  )
)
contrast_matrix_expanded <- data.frame(
  contrast_names = c(), 
  numerators = c(),
  denominators = c()
)
for (n in 1:nrow(contrast_matrix_formula)) {
  contrast_name <- contrast_matrix_formula$contrast_names[n]
  numerator <- contrast_matrix_formula$numerators[n]
  denominator <- contrast_matrix_formula$denominators[n]
  for (kit in kit_order_3p) {
    numerator1 <- gsub('{{KIT}}', kit, numerator, fixed = TRUE)
    denominator1 <- gsub('{{KIT}}', kit, denominator, fixed = TRUE)
    if (kit == 'Flex') {
      ## flex is reference level, doesn't have extra terms
      terms <- unlist(str_extract_all(numerator1, '(?:\\+ \\(1\\/6\\)\\s*)*[+*] KitFlex\\S*\\s*'))
      for (term in terms) {
        if (!gsub('\\s*\\+\\s*', '', term) %in% make.names(colnames(design_matrix))) {
          numerator1 <- sub(term, '', numerator1, fixed = TRUE)
        }
      }
      terms <- unlist(str_extract_all(denominator1, '(?:\\+ \\(1\\/6\\)\\s*)*[+\\*] KitFlex\\S*\\s*'))
      for (term in terms) {
        if (!gsub('\\s*\\+\\s*', '', term) %in% make.names(colnames(design_matrix))) {
          denominator1 <- gsub(term, '', denominator1, fixed = TRUE)
        }
      }
      
      # numerator1 <- gsub('(?:\\+ \\(1\\/6\\)\\s*)*[+*] KitFlex\\S*', '', numerator1)
      # denominator1 <- gsub('(?:\\+ \\(1\\/6\\)\\s*)*[+*] KitFlex\\S*', '', denominator1)
    } 
    out <- data.frame(contrast_names = paste0(kit, '.', contrast_name),
                      numerators = numerator1,
                      denominators = denominator1)
    contrast_matrix_expanded <- rbind(contrast_matrix_expanded, out)
  }
}
contrast_matrix_expanded
```

```{r}
## have to make valid names first
## Current workflow requires user to make valid names instead
contrast_matrix <- contrast_matrix_expanded

## Then make contrasts
contrast_matrix <- apply(contrast_matrix, MARGIN = 1, function(row) {
  paste0('(', row[2], ') - (', row[3], ')')
})
contrast_matrix <- limma::makeContrasts(contrasts = contrast_matrix,
    levels = make.names(colnames(design_matrix)))
colnames(contrast_matrix) <- contrast_matrix_expanded$contrast_names
View(contrast_matrix)
```

## Extract

```{r, eval = FALSE}
model_results <- lapply(objs, function(x) {
  kit <- as.character(unique(x$Kit))
  out <- apply(contrast_matrix, 2, results, object = x, alpha = 0.05)
  names(out) <- paste0(kit, '.', colnames(contrast_matrix))
  out
})
model_results <- unlist(unname(model_results))
# model_results <- apply(contrast_matrix, 2, results, object = pb_obj, alpha = 0.05)
# tmp <- results(pb_obj, contrast = contrast_matrix[,1])
# names(model_results) <- colnames(contrast_matrix)
saveRDS(model_results, here('rds/3p/dge_model_results_splitkit_downsampled_coarse_celltypes.rds'))
```

```{r}
model_results <- readRDS(here('rds/3p/dge_model_results_splitkit_downsampled_coarse_celltypes.rds'))
```

# Vis DGE counts

```{r}
dge_up <- lapply(model_results, function(x){
  rownames(filter(as.data.frame(x), log2FoldChange > 0 & padj < 0.05))
})
dge_down <- lapply(model_results, function(x){
  rownames(filter(as.data.frame(x), log2FoldChange < 0 & padj < 0.05))
})
dge_joint <- lapply(setNames(names(dge_down), names(dge_down)), function(x) {
  c(paste0(dge_up[[x]], '+'), paste0(dge_down[[x]],'-'))
})
```

## upset

```{r}

kits <- levels(metadata$Kit)
intersections <- lapply(1:length(kits), function(x) kits[-x])
intersections <- c(
  levels(metadata$Kit),
  list(list('Fluent_v4', 'Fluent_V')),
  list(list('GEMX3P', 'NextGEM3P')),
  list(list('Flex', 'GEMX3P', 'NextGEM3P')),
  list(list('Parse_v3', 'Scale')),
  list(list('Flex', 'NextGEM3P', 'GEMX3P', 'Fluent_v4', 'Fluent_V')),
  list(list('Flex', 'NextGEM3P', 'GEMX3P', 'Scale', 'Parse_v3')),
  list(list('Fluent_v4', 'Fluent_V', 'Scale', 'Parse_v3')),
  intersections,
  # list(kits),
  list(levels(metadata$Kit))
)
```

```{r}
for (contrast in contrast_matrix_formula$contrast_names) {
  
  plotdata <- dge_joint[grepl(contrast, names(dge_joint))]
  names(plotdata) <- lapply(names(plotdata),
                            function(x) {(strsplit(x, split='.', fixed=TRUE) |>
                                            unlist())[[1]] }) |>
    unlist() 
  plotdata <- fromList(plotdata)
  upset(plotdata, 
        sets=levels(metadata$Kit),
        # nsets=8,
        # nintersects = 140, 
        keep.order = TRUE,
        intersections = intersections,
        order.by='degree',
        # cutoff=100,
        decreasing = FALSE,
        mainbar.y.label='Gene count',
        number.angles = 0,
        empty.intersections=FALSE) ->
    figures[[paste0('upset_', contrast)]]
}
figures[['upset_B.vs.all']]
```

## Euler

```{r}
for (contrast in contrast_matrix_formula$contrast_names) {
  plotdata_up <- dge_joint[grepl(contrast, names(dge_joint))]
  names(plotdata) <- lapply(names(plotdata),
                            function(x) {(strsplit(x, split='.', fixed=TRUE) |>
                                            unlist())[[1]] }) |>
    unlist()
  figures[[paste0('euler_', contrast)]] <- plot(eulerr::euler(plotdata, shape='ellipse'))
}
figures[['euler_B.vs.T']]
```

## Correlation

```{r, fig.width=16, fig.height=15}
library(ggpubr)
for (contrast in contrast_matrix_formula$contrast_names) {
  plotdata <- model_results[grepl(contrast, names(model_results))]
  plotdata <- lapply(plotdata, function(x) {
    x |> 
      as.data.frame() |>
      rownames_to_column('gene') |>
      # filter(padj < 0.05) |>
      select(gene, stat, padj, log2FoldChange)
  }) |> rbindlist(idcol='Kit') 
  
  plotdata <- plotdata |> 
    inner_join(plotdata, by = "gene", relationship = 'many-to-many') |>
    mutate(Comparison = paste(Kit.x, "vs", Kit.y)) |>
    filter(padj.x < 0.05) |>
    mutate(Kit.x = gsub('([^\\.]+)\\..+', '\\1', Kit.x),
           Kit.y = gsub('([^\\.]+)\\..+', '\\1', Kit.y)) 
  
  ggplot(plotdata, aes(x=log2FoldChange.x, y=log2FoldChange.y)) +
    geom_bin2d(bins=70) +
    geom_abline(slope=1, lty='dashed', color='black', alpha=0.4) +
    geom_smooth(method='lm') +
    stat_cor(aes(label = after_stat(rr.label))) + 
    facet_grid(Kit.y ~ Kit.x, drop = TRUE, 
               labeller = labeller(Kit.x = label_function(), Kit.y = label_function())) ->
    figures[[paste0('corr_pairplot_', contrast)]]
}
figures[[paste0('corr_pairplot_', contrast)]]
```

## Barchart

```{r}
for (cont in contrast_matrix_formula$contrast_names) {
  plotdata_up <- lapply(dge_up[grepl(cont, names(dge_up))], length)
  names(plotdata_up) <- lapply(names(plotdata_up),
                            function(x) {(strsplit(x, split='.', fixed=TRUE) |>
                                            unlist())[[1]] }) |>
  unlist()
  plotdata_up <- plotdata_up|>
    melt() |>
    mutate(direction = 'Up')

  plotdata_down <- lapply(dge_down[grepl(cont, names(dge_down))], length)
  names(plotdata_down) <- lapply(names(plotdata_down),
                            function(x) {(strsplit(x, split='.', fixed=TRUE) |>
                                            unlist())[[1]] }) |>
  unlist()
  plotdata_down <- plotdata_down|>
    melt() |>
    mutate(direction = 'Down', value = -1 * value)
  plotdata <- rbind(plotdata_up, plotdata_down) |>
    mutate(Kit = factor(L1, kit_order_3p))
  figures[['dge_barchart']][[cont]] <-
    ggplot(plotdata, aes(x=Kit, y=value, group=direction)) +
    geom_bar(aes(y = value, fill = direction), stat = "identity", position = "identity") +
    # geom_bar(subset = .(direction == "Down"), 
    #   aes(y = -value, fill = direction), stat = "identity", position = "dodge") + 
    scale_x_discrete(labels = label_function()) +
    geom_hline(yintercept = 0,colour = "grey90") +
    lims(y=c(-max(abs(plotdata$value)),max(abs(plotdata$value)))) +
    labs(x='Kit', y='Significant DEG', fill = 'Direction\nof DGE', title=cont)
}
```

```{r, fig.width=8}
figures[['dge_barchart']]
```


# Volcano plots

```{r}
generateVolcanoPlot <- function(result,
                                labels = 20,
                                FCcutoff = log2(1.3),
                                pCutoff = 0.05,
                                title = NULL,
                                caption = NULL,
                                subtitle = NULL,
                                ...) {
  volData <- result[!is.na(result$padj), ]
  if (is.numeric(labels)) {
    volData <- volData[order(volData$padj), ]
    labels <- rownames(volData[1:labels, ])
  }
  volplot <- EnhancedVolcano::EnhancedVolcano(volData,
    x = "log2FoldChange",
    y = "padj",
    lab = rownames(volData),
    selectLab = labels,
    drawConnectors = TRUE,
    colConnectors = "lightgrey",
    pCutoff = pCutoff,
    FCcutoff = FCcutoff,
    title = title,
    caption = caption,
    subtitle = subtitle,
    ...
  )
  return(volplot)
}
```

```{r, warning=FALSE}
figures[['dge_volcano']] <- lapply(model_results, generateVolcanoPlot, xlim = c(-10,10))
figures[['dge_volcano']]$Flex.B.vs.T
figures[['dge_volcano']]$Scale.B.vs.T
```


# Celltype markers

```{r}
celltype_markers <- read.csv('/fh/fast/_IRC/FHIL/grp/FHIL_knowledgebase/biology/celltype_markers.csv')
# shared_genes <- read.table(here('rds/3p/gene_whitelist.txt'))$V1

parse_marker_table <- function(celltype_dataframe) {
  celltype_dataframe %>%
    filter(tissue == 'PBMC') %>%
    # filter(expression_level == 'Increased') |>
    filter(confidence %in% c('high', 'med')) #%>%
    # filter(celltype %in% c('T', 'CD4+ T', 'CD8+ T',
    #                        'B', 'B naive', 'B memory',
    #                        'Monocyte', 'Non-classical monocyte', 'Classical monocyte',
    #                        'Erythrocyte', 'Granulocyte',
    #                        'NK', 'Dendritic', 'Megakaryocyte', 'pDC')) %>%
    # select(gene_symbol, celltype) %>%
    # filter(gene_symbol %in% names(pb_obj@rowRanges)) 
}
celltype_markers <- parse_marker_table(celltype_markers)
celltype_markers_cd8.cd4 <- celltype_markers |> 
  filter((celltype == 'CD4+ T' & relative_to == 'CD8+ T') |
           (celltype == 'CD8+ T' & relative_to == 'CD4+ T'))

celltype_markers_Bnaive.v.Bmem <- celltype_markers |> 
  filter((celltype == 'B naive' & relative_to == 'B memory') |
           (celltype == 'B memory' & relative_to == 'B naive'))

celltype_markers_Bmem.v.CD4T <- celltype_markers |> 
  filter((celltype == 'CD4+ T' & expression_level == 'Increased') |
           (celltype == 'B memory' & expression_level == 'Increased'))

celltype_markers_B.v.T <- celltype_markers |> 
  filter((celltype == 'T') |
           (celltype == 'B'))

# celltype_markers <- parse_marker_table(celltype_markers)
# 
# # flex_genes <- Reduce(intersect, list(Features(objs$Flex_F1A),
# #                         Features(objs$Flex_F1B),
# #                         Features(objs$Flex_F5A),
# #                         Features(objs$Flex_F5B)))
# # celltype_markers_modules <- lapply(celltype_markers, function(x)
# #   x[x %in% shared_genes & x %in% names(pb_obj@rowRanges)]
# # )
# celltype_markers_modules$`B memory` <- 
#   c(celltype_markers_modules$`B memory`, celltype_markers_modules$B)
# celltype_markers_modules$`B naive` <- 
#   c(celltype_markers_modules$`B naive`, celltype_markers_modules$B)
# 
# celltype_markers_modules$`CD4+ T` <- 
#   c(celltype_markers_modules$`CD4+ T`, celltype_markers_modules$`T`)
# celltype_markers_modules$`CD8+ T` <- 
#   c(celltype_markers_modules$`CD8+ T`, celltype_markers_modules$`T`)
# 
# celltype_markers_modules$`Classical monocyte` <- 
#   c(celltype_markers_modules$`Classical monocyte`, celltype_markers_modules$`Monocyte`)
# celltype_markers_modules$`Non-classical monocyte` <- 
#   c(celltype_markers_modules$`Non-classical monocyte`, celltype_markers_modules$`Monocyte`)
# 
# celltype_markers_modules <- lapply(celltype_markers_modules, unique)
# celltype_markers_shared <- filter_marker_genes(celltype_markers, shared_genes)
```

## B.v.T

```{r, fig.width=12, fig.height=5}
plotdata <- model_results[grepl('B.vs.T', names(model_results))]
names(plotdata) <- lapply(names(plotdata),
                          function(x) {(strsplit(x, split='.', fixed=TRUE) |>
                                          unlist())[[1]] }) |>
  unlist()
plotdata <- lapply(plotdata, function(x) {
  x |> 
    as.data.frame() |>
    rownames_to_column('gene')
}) |>
  rbindlist(idcol = 'kit')
plotdata <- plotdata |>
  filter(gene %in% celltype_markers_Bmem.v.CD4T$gene_symbol) 
plotdata |> 
  ggplot(aes(x=padj, y=log2FoldChange, color = kit)) +
  geom_point(alpha=0.8) +
  scale_color_manual(values = unlist(color_palette$kits), labels = label_function(), na.translate = FALSE) +
  geom_vline(xintercept = 0.05, linetype='dashed') +
  # scale_x_continuous(transform = 'log2') +
  facet_wrap(~ gene, nrow=2) ->
  figures[['markers_bvst']]
figures[['markers_bvst']]
```

# Significant gene correlation

```{r}
dge_up <- lapply(model_results, function(x){
  filter(as.data.frame(x), log2FoldChange > 0 & padj < 0.05) |>
    rownames_to_column('gene')
})
# names(dge_up) <- lapply(names(dge_up),
#                           function(x) {(strsplit(x, split='.', fixed=TRUE) |>
#                                           unlist())[[1]] }) |>
#   unlist()
dge_down <- lapply(model_results, function(x){
  filter(as.data.frame(x), log2FoldChange < 0 & padj < 0.05) |>
    rownames_to_column('gene')
})
# names(dge_down) <- lapply(names(dge_down),
#                           function(x) {(strsplit(x, split='.', fixed=TRUE) |>
#                                           unlist())[[1]] }) |>
#   unlist()
```

## B vs T

```{r, fig.width=10, fig.height=10}
plotdata <- rbindlist(dge_up[grepl('B.vs.T', names(dge_up))], idcol = 'Kit') |>
  mutate(Kit = sub('([^\\.]+)\\..+', '\\1', Kit)) |>
  select(Kit, gene, stat) |>
  # tidyr::pivot_wider(names_from = Kit, values_from=stat) |> 
  # melt(variable.name = 'Kit', value.name = 'stat') |>
  mutate(Kit = factor(Kit, levels = kit_order_3p))
plotdata <- plotdata |> 
    inner_join(plotdata, by = "gene") |>
  filter(Kit.x != Kit.y) %>%  # Remove comparisons of the same kit
  mutate(Comparison = paste(Kit.x, "vs", Kit.y))
# kit_pairs <- combn(names(plotdata)[-1], 2, simplify = FALSE)
# df_long <- bind_rows(lapply(kit_pairs, function(pair) {
#   plotdata %>%
#     select(gene, all_of(pair)) %>%
#     dplyr::rename(X = pair[1], Y = pair[2]) %>%
#     mutate(Comparison = paste(pair[1], "vs", pair[2]))  # Create facet label
# }))

plotdata |>
  ggplot(aes(x=stat.x, y=stat.y)) +
  geom_point() +
  geom_abline(slope = 1, intercept = 0, linetype = "dashed", color = "black") +
  facet_grid(Kit.x ~ Kit.y, labeller = labeller(Kit.y = label_function(), Kit.x = label_function())) +
  geom_smooth(method = 'lm') ->
  figures[['sig_gene_correlation_bvst']]
figures[['sig_gene_correlation_bvst']]

```

# GSEA

```{r}
pathways <- msigdbr(species = "human", category = "C2") |>
  filter(gs_subcat  == 'CP:BIOCARTA')
pathways <- pathways %>% dplyr::select(gene_symbol, gs_name) %>% unstack()
```

```{r, eval=FALSE}

gsea_results <- list()
for (result in names(model_results)){
  ranked_genes <- model_results[[result]] %>% 
    as.data.frame() |>
    rownames_to_column('gene') |>
    filter(!is.na(padj))
  ranked_genes$stat <- ranked_genes$stat#ranked_genes$logFC * -log10(ranked_genes$padj + .Machine$double.xmin)
  ranked_genes <- setNames(ranked_genes$stat, nm = ranked_genes$gene)
  ranked_genes <- sort(ranked_genes)
  res <- fgsea::fgseaSimple(pathways, ranked_genes, nperm = 5000) |>
    mutate(name = gsub('BIOCARTA_(.+)_PATHWAY', '\\1', pathway))
  gsea_results[[result]] <- res
}

# Save GSEA results
saveRDS(gsea_results, here('rds/3p/gsea_results.rds'))
```

```{r}
gsea_results <- readRDS(here('rds/3p/gsea_results.rds'))
```

```{r}
gsea_results$Flex.T.vs.M |>
  arrange(pval)
gsea_results$Parse_v3.B.vs.T |>
  arrange(pval)
## |>
#   filter(pathway %in% c(
#   'BIOCARTA_BLYMPHOCYTE_PATHWAY',
#   'BIOCARTA_BCR_PATHWAY',
#   'BIOCARTA_TCR_PATHWAY',
#   'BIOCARTA_MONOCYTE_PATHWAY'))
```

## euler

```{r}
for (contrast in contrast_matrix_formula$contrast_names) {
  plotdata <- gsea_results[grepl(contrast, names(gsea_results))]
  plotdata <- lapply(plotdata, function(x){
    x <- x |>
      filter(padj <= 0.05) |>
      mutate(x = paste0(name, ifelse(ES < 0, '-', '+')))
    x$x
  })
  # names(plotdata) <- lapply(names(plotdata),
  #                           function(x) {(strsplit(x, split='.', fixed=TRUE) |>
  #                                           unlist())[[1]] }) |>
  #   unlist()
  figures[[paste0('euler_gsea_', contrast)]] <- plot(eulerr::euler(plotdata, shape='ellipse'))
}
figures[['euler_gsea_T.vs.M']]
```

## upset

```{r}
for (contrast in contrast_matrix_formula$contrast_names) {
  
  plotdata <- gsea_results[grepl(contrast, names(gsea_results))]
  plotdata <- lapply(plotdata, function(x){
    x <- x |>
      filter(padj <= 0.05) |>
      mutate(x = paste0(name, ifelse(ES < 0, '-', '+')))
    x$x
  })
  names(plotdata) <- lapply(names(plotdata),
                            function(x) {(strsplit(x, split='.', fixed=TRUE) |>
                                            unlist())[[1]] }) |>
    unlist() 
  plotdata <- fromList(plotdata)
  upset(plotdata, 
        sets=levels(metadata$Kit),
        # nsets=8,
        # nintersects = 140, 
        keep.order = TRUE,
        intersections = intersections,
        order.by='degree',
        # cutoff=100,
        decreasing = FALSE,
        mainbar.y.label='Gene count',
        number.angles = 0,
        empty.intersections=FALSE) ->
    figures[[paste0('upset_gsea_', contrast)]]
}
figures[['upset_gsea_T.vs.M']]
```


## Volcano

```{r}

gseaVolcanoPlot <- function(result,
                                labels = 10,
                                NEScutoff = 0.5,
                                pCutoff = 0.1,
                                title = NULL,
                                caption = NULL,
                                subtitle = NULL,
                                ...) {
  volData <- result[!is.na(result$padj), ] |> 
      arrange(padj)
  if (is.numeric(labels)) {
    # tmp <- volData |> filter(padj < pCutoff) 
    labels <- volData$name[1:labels]
  }
  volplot <- EnhancedVolcano::EnhancedVolcano(volData,
    x = "NES",
    y = "padj",
    lab = volData$name,
    selectLab = labels,
    drawConnectors = TRUE,
    colConnectors = "lightgrey",
    pCutoff = pCutoff,
    FCcutoff = NEScutoff,
    title = title,
    caption = caption,
    subtitle = subtitle,
    max.overlaps = 100,
    xlim = c(-3,3),
    ylim = c(0, 3),
    ...
  )
  return(volplot)
}
```

```{r}
figures[['gsea_volcano']] <- lapply(gsea_results, gseaVolcanoPlot)
figures[['gsea_volcano']]$Flex.B.vs.T
figures[['gsea_volcano']]$Flex.T.vs.M
figures[['gsea_volcano']]$Flex.M.vs.B
```

## GSEA dotplot

```{r, fig.width=12, fig.height=6}
pathways_of_interest <- c('TID', 'TCR', 'TCRA', 'TCYTOTOXIC', 'CTL', 'LYM', 'LYMPHOCYTE', 'MONOCYTE', 'BCR', 'BLYMPHOCYTE')
plotdata <- rbindlist(gsea_results, idcol='contrast') |>
  mutate(kit = case_when(
    grepl('Flex', contrast) ~ 'Flex',
    grepl('NextGEM3P', contrast) ~ 'NextGEM3P',
    grepl('GEMX3P', contrast) ~ 'GEMX3P',
    grepl('Fluent_v4', contrast) ~ 'Fluent_v4',
    grepl('Fluent_V', contrast) ~ 'Fluent_V',
    grepl('Parse_v3', contrast) ~ 'Parse_v3',
    grepl('Scale', contrast) ~ 'Scale',
  ), comp = case_when(
    grepl('B.vs.T', contrast) ~ 'B.vs.T',
    grepl('T.vs.M', contrast) ~ 'T.vs.M',
    grepl('M.vs.B', contrast) ~ 'M.vs.B',
    grepl('B.vs.all', contrast) ~ 'B',
    grepl('T.vs.all', contrast) ~ 'T',
    grepl('M.vs.all', contrast) ~ 'M',
    # grepl('CD8.vs.CD4', contrast) ~ 'CD8.vs.CD4',
    .default = NA
  )) |>
  filter(!is.na(comp), name %in% pathways_of_interest) |>
  mutate(name = factor(name, levels = pathways_of_interest),
         sig = case_when(
           padj < 0.0005 ~ '***',
           padj < 0.005 ~ '**',
           padj < 0.05 ~ '*',
           .default = NA
         ))
ggplot(plotdata, aes(x=name, y=1, color=NES, size=-log(padj), label = sig)) +
  geom_point() +
  facet_grid(comp ~ kit) +
    scale_color_gradient2(low="blue",
                          mid="white",
                          high="red",
                          midpoint=0,
                          breaks=c(-2,-1,0,1,2),
                          limits=c(min(plotdata$NES,-2),
                                   max(plotdata$NES,2))) +
    # theme_classic(11) +
    theme(panel.grid.major = element_line(colour = "grey92"),
          panel.grid.minor = element_line(colour = "grey92"),
          panel.grid.major.y = element_line(colour = "grey92"),#element_blank(),
          panel.grid.minor.y = element_line(colour = "grey92")) +
    theme(axis.text.x = element_text(angle = 45, 
                                     # vjust = 0.5, 
                                     hjust=1),
          axis.text.y = element_blank(),
          axis.ticks.y = element_blank()) +
    labs(x="Gene set", y=NULL,
         color = "Normalized\nenrichment\nscore",
         size="Adjusted p-value",
         caption = '* indicates adjusted p-value <= 0.05') +
   geom_text(na.rm = TRUE, color = 'white', size = 3) + 
  scale_size_area(max_size = 6, breaks=c(1,2,3), labels = c(0.1, 0.01, 0.001)) ->
  figures[['gsea_dotplot']]
figures[['gsea_dotplot']]
  
```

```{r}
gseaDotplot_joint <- function(result,
                              pathway_order = NULL,
                              x_order = NULL,
                              significance = c(0.05, 0.01, 0.001),
                              breaks = c(0.1,0.01,0.001,0.0001),
                              cap_pvalues = TRUE,
                              p_val_col = 'pval',
                              use_shortened_pathway_names = FALSE
                              ){
  if (use_shortened_pathway_names){
    result$pathway <- result$pathway_short
  }
  if (!is.null(pathway_order)) {
    if (all(pathway_order %in% unique(result$pathway))){
      pathway_order <- order(factor(result$pathway, levels = pathway_order))
      result$pathway <- .wrap_underscore_strings_balance(result$pathway,36)
      ## reordering
      result <- result[pathway_order,]
      result$pathway <- factor(result$pathway, levels = unique(result$pathway))
    } else {
      warning('pathways specified in pathway_order not found, defaulting to arbitrary order')
    }
  } else {
    result$pathway <- .wrap_underscore_strings_balance(result$pathway,36)
  }
  
  if (!is.null(x_order)) {
    if (all(x_order %in% unique(result$ID))){
      result$ID <- factor(result$ID, levels = x_order)
    } else {
      warning('Specified x_order not all found in data, defaulting to arbitrary order')
    }
  }
  
  if (cap_pvalues) {
    ## Set the minimum value to be 1/10th of the smallest label
    cap_max = tail(breaks, 1)/10
    result[[p_val_col]] <- pmax(cap_max, result[[p_val_col]])
  }
  range <- c(ceiling(-log(max(result[[p_val_col]]))),
             floor(-log(min(result[[p_val_col]]))))+1
  result$label <- NA
  caption <- ''
  if (!is.null(significance)) {
    if (is.numeric(significance)) {
      label <- '*'
      for (cutoff in significance) {
        result$label <- ifelse(result[[p_val_col]] < cutoff, label, result$label)
        caption <- paste(caption, label, '<', cutoff, ';', sep = ' ')
        label <- paste0(label, '*')
      }
      # result$label[is.numeric(result$label)] <- NA
    } else {
      stop('Significance argument should be a numeric vector')
    }
  } 
  ggplot(result, aes(x=.data$ID, y=.data$pathway, size=.data[[p_val_col]],
                           color=.data$NES, label = .data$label)) + 
    geom_point() +
    scale_color_gradient2(low="blue",
                          mid="white",
                          high="red",
                          midpoint=0,
                          breaks=c(-2,-1,0,1,2),
                          limits=c(min(result$NES,-1),
                                   max(result$NES,1))) +
    theme_classic(11) +
    theme(panel.grid.major = element_line(colour = "grey92"),
          panel.grid.minor = element_line(colour = "grey92"),
          panel.grid.major.y = element_line(colour = "grey92"),#element_blank(),
          panel.grid.minor.y = element_line(colour = "grey92")) +
    theme(axis.text.x = element_text(angle = 90, 
                                     vjust = 0.5, 
                                     hjust=1)) +
    labs(x="Comparison",
         y="Gene set", 
         color = "Normalized\nenrichment\nscore",
         size="p-value",
         title="GSEA pathway enrichments",
         caption = caption) +
    scale_radius(range=range,
                 trans=reverselog_trans(),
                 breaks=breaks
                 ) +
    geom_text(na.rm = TRUE, color = 'white', size = 3)
}
```



# Render report

```{r, eval=TRUE}
## This only works if all the variables called in the format file 
## are in the current environment. Check the format file to tweak
## aesthetics 
rmarkdown::render(here('processing_scripts/3p/09-dge.format.Rmd'),
                  output_file = '09-dge_splitmodel.html',
                  output_dir = here('reports/3p/'))
```
